refactor: move BasicAsyncTask to utils folder
